Move undulator_dcm
